move the PartialEq implementation for Node into tests.rs, which is the only place it's used. It's implemented recurisvely so it's not safe for production use
